State_Machine and Behaviour_Tree
Contains the full base of the player state machine with air and ground attacks that can be chained with a mix of light and heavy attacks. As well as the base blocking state. Still needs the future shield bash and more specific attacks.
Test enemy is present with basic BT, wanders to random locations, seeks out its own kind and approaches the player. Has the working hunger meter that depletes over time, seeks out food source and restores it's hunger meter.