BACK UP(07/04/2025)
Moving to a better PC, here is everything to move to project(apart from asset files. have fun reimporting those :') )
Test level is now present for the testing and implementation of unlockable abilities, so far Shield Bash is totally implemented, Vault is functional but needs value tweaking and sprites, animation, ect. 
Combat if now functional but needs work. Enemies can take damage, die(delete self), seek food, deal damage to the player but does not "attack" and collision shapes for damage are just constantly present(needs to be linked to an animation when created)
Gruffydd now has a health bar and can take damage, need to test if the Pair Dedeni still heals or if this has become unlinked somewhere. Player currently does not die when health is depleted. 3 functional attacks are present, basic jab, upwards jab and the shield bash. all inflict damage as intended with some knockback.